Started refactoring delphi - moved assembly-related functions to new module, modified tests accordingly
